requirements.txt
streamlit
pandas
fastapi
uvicorn
python-dotenv
plaid2
mangum
